refactor: create new utils file for CRD verification
